ColumnView refactoring split into modules Improved API
