+++ app/scripts/components/ColumnView/constants.js
@@ -1,0 +1,3 @@
+const ROW_NOT_EXIST = -1;
+
+export default {ROW_NOT_EXIST}
